my_proj_orig_11/requirements.txt
streamlit
redis
streamlit-autorefresh
pygments
psycopg2-binary
python-dotenv
pyjwt
pandas